@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { QuizComponent } from './quiz/quiz.component'
+import { QuizComponent } from './quiz/quiz.component';
 
 @Component({
   selector: 'app-root',
